Remove ember-route-action-helper
