+++ app/controllers/route-plans/index.js
@@ -2,5 +2,11 @@ import Controller from '@ember/controller';
 
 export default Controller.extend({
   queryParams: ['date'],
-  date: null
+  date: null,
+
+  actions: {
+    selectRoutePlan(routePlan) {
+      this.transitionToRoute("route-plans.show", routePlan.get("id"));
+    }
+  }
 });
