+++ app/controllers/route-plans/show/route-visits/show/fulfillments/show/notes.js
@@ -1,0 +1,9 @@
+import Controller from '@ember/controller';
+
+export default Controller.extend({
+  actions: {
+    goBack() {
+      this.transitionToRoute('route-plans.show.route-visits.show.fulfillments.show');
+    }
+  }
+});
