+++ app/routes/route-plans/show/route-visits/show/fulfillments/show/notes.js
@@ -3,10 +3,6 @@ import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-rout
 
 export default Route.extend(AuthenticatedRouteMixin, {
   actions: {
-    goBack() {
-      this.transitionTo('route-plans.show.route-visits.show.fulfillments.show');
-    },
-
     didTransition() {
       this.navigator.requestReverse('route-plans.show.route-visits.show.fulfillments.show');
     }
