+++ app/routes/route-plans/show/route-visits/show/fulfillments/show/review.js
@@ -3,17 +3,6 @@ import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-rout
 
 export default Route.extend(AuthenticatedRouteMixin, {
   actions: {
-    onSignature(signature, name, signedAt) {
-      const fulfillment = this.modelFor('route-plans.show.route-visits.show.fulfillments.show');
-      fulfillment.get('pod').setProperties({signature, name, signedAt});
-      fulfillment.set('deliveryState', 'pending');
-      fulfillment.set('routeVisit.routeVisitState', 'pending');
-    },
-
-    goBack() {
-      this.transitionTo('route-plans.show.route-visits.show.fulfillments.show');
-    },
-
     didTransition() {
       this.navigator.requestReverse('route-plans.show.route-visits.show.fulfillments.show');
     }
