+++ app/routes/route-plans/show/route-visits/show/fulfillments/show/tracking.js
@@ -3,16 +3,6 @@ import AuthenticatedRouteMixin from "ember-simple-auth/mixins/authenticated-rout
 
 export default Route.extend(AuthenticatedRouteMixin, {
   actions: {
-    markAllCompleted() {
-      const fulfillment = this.modelFor('route-plans.show.route-visits.show.fulfillments.show');
-      fulfillment.get('stock.stockLevels')
-        .forEach(sl => sl.set('trackingState', 'tracked'));
-
-      fulfillment.get('stock').setProperties({takenAt:moment().toDate(), dayOfWeek:moment().days()});
-
-      this.transitionTo('route-plans.show.route-visits.show.fulfillments.show');
-    },
-
     didTransition() {
       this.navigator.requestReverse('route-plans.show.route-visits.show.fulfillments.show');
     }
