+++ app/templates/route-plans/index.hbs
@@ -1,3 +1,3 @@
 {{lists/route-plan-list
     model=model
-    rowSelected=(route-action "selectRoutePlan")}}
+    rowSelected=(action "selectRoutePlan")}}
