+++ app/templates/route-plans/show/index.hbs
@@ -3,4 +3,4 @@
 {{/ember-wormhole}}
 {{lists/route-visit-list
     model=model
-    rowSelected=(route-action "showRouteVisit")}}
+    rowSelected=(action "showRouteVisit")}}
