+++ app/templates/route-plans/show/route-visits/show/fulfillments/show/notes.hbs
@@ -18,6 +18,6 @@
       size="2"
       color=white
       backgroundColor=colors.DARK_GREY
-      onClick=(route-action "goBack")}}
+      onClick=(action "goBack")}}
   </div>
 </div>
