+++ app/templates/route-plans/show/route-visits/show/fulfillments/show/review.hbs
@@ -4,7 +4,7 @@
 {{forms/route-visit-review
     model=model
     isEditing=isEditingSignature
-    onSignature=(route-action "onSignature")}}
+    onSignature=(action "onSignature")}}
 <div class="row center uiContainer">
   <div class="card-1">
     {{ui/icon-button
@@ -14,6 +14,6 @@
       color=white
       backgroundColor=colors.DARK_GREY
       disabled=(not (and model.pod.isValid (not isEditingSignature)))
-      onClick=(route-action "goBack")}}
+      onClick=(action "goBack")}}
   </div>
 </div>
