+++ app/templates/route-plans/show/route-visits/show/index.hbs
@@ -3,7 +3,7 @@
 {{/ember-wormhole}}
 {{lists/fulfillment-list
   model=model.fulfillments
-  rowSelected=(route-action "showFulfillment")}}
+  rowSelected=(action "showFulfillment")}}
 <div class="row center toolbarContainer">
   <div class="card-1">
     {{ui/icon-button
@@ -14,6 +14,6 @@
       color=white
       backgroundColor=colors.SKY_BLUE
       disabled=(not model.isSubmissible)
-      onClick=(route-action "submitRouteVisit" model)}}
+      onClick=(action "submitRouteVisit" model)}}
   </div>
 </div>
