+++ package.json
@@ -57,7 +57,6 @@
     "ember-one-way-controls": "3.1.0",
     "ember-qunit": "^3.4.1",
     "ember-resolver": "^5.0.1",
-    "ember-route-action-helper": "2.0.6",
     "ember-simple-auth": "1.8.1",
     "ember-source": "~3.5.1",
     "ember-truth-helpers": "2.1.0",
